requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.1
dnspython==2.4.2
docopt==0.6.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.104.0
frozenlist==1.4.0
gevent==23.9.1
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
greenlet==3.0.0
h11==0.14.0
hbmqtt==0.9.6
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.320
langsmith==0.0.49
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
orjson==3.9.9
packaging==23.2
paho-mqtt==1.6.1
passlib==1.7.4
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyparsing==3.1.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tenacity==8.2.3
tqdm==4.66.1
transitions==0.2.5
typing-inspect==0.9.0
typing_extensions==4.8.0
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
yarl==1.9.2
zope.event==5.0
zope.interface==6.1
